refactor: header and body components trim
