+++ package.json
@@ -46,6 +46,7 @@
     "redux-devtools-extension": "^2.13.9",
     "redux-saga": "^1.1.3",
     "regenerator-runtime": "^0.13.7",
+    "slick-carousel": "^1.8.1",
     "utility-types": "^3.10.0",
     "yup": "^0.29.3"
   },
