--- src/components/headerComp/detailButton.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const DetailClick = () => {
-  return <></>;
-};
