--- src/components/headerComp/headerRightSide.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const HeaderRightSide = () => {
-  return (
-    <>
-      <span className="rightHover">
-        <span>쿠폰</span> | <span>로그인</span> | <span>회원가입</span>
-      </span>
-    </>
-  );
-};
-
-export default HeaderRightSide;
